Fixed casting error and removed loop-iteration for Skript 2.8+ as they added that in

+++ src/main/java/me/sashie/skriptyaml/utils/versions/V2_3.java
@@ -77,9 +77,9 @@ public class V2_3 implements SkriptAdapter {
 	@Override
 	public long getTime(Date date) {
 		try {
-			return (int) getTimeStampMethod.invoke(date);
+			return (long) getTimeStampMethod.invoke(date);
 		} catch (InvocationTargetException | IllegalAccessException e) {
-			return 0;
+			return 0L;
 		}
 	}
